All existing interfaces folders and files were renamed in one manner
