--- src/types/IBreadCrumbsItem.ts
@@ -1,9 +1,0 @@
-export default interface IBreadCrumbsItem {
-  crumbNum: number;
-  title: string;
-  disabled: boolean;
-  to: {
-    path: string;
-  };
-  crumbIconName?: string;
-}
