--- src/types/IDepartmentData.ts
@@ -1,4 +1,0 @@
-export default interface IDepartmentData {
-  id: number;
-  name: string;
-}
