--- src/types/IForm.ts
@@ -1,4 +1,0 @@
-export interface IForm {
-  email: string;
-  password: string;
-}
