--- src/types/INavigationItem.ts
@@ -1,5 +1,0 @@
-export interface INavigationItem {
-  title: string;
-  icon: string;
-  link: string;
-}
