--- src/types/IPositionData.ts
@@ -1,4 +1,0 @@
-export default interface IPositionData {
-  id: number;
-  name: string;
-}
