+++ src/types/departmentData.ts
@@ -1,0 +1,4 @@
+export interface IDepartmentData {
+  id: number;
+  name: string;
+}
