+++ src/types/form.ts
@@ -1,3 +1,8 @@
+export interface IForm {
+  email: string;
+  password: string;
+}
+
 export interface IAuthForm {
   title: string;
   buttonText: string;
