+++ src/types/positionData.ts
@@ -1,0 +1,4 @@
+export interface IPositionData {
+  id: number;
+  name: string;
+}
