+++ src/types/userSkillsUI.ts
@@ -6,7 +6,14 @@ export enum Mastery {
   Expert,
 }
 
-export interface ISkillMastery {
+export interface ISkill {
+  id: number;
+  createdAt: string;
+  name: string;
+  category: string;
+}
+
+export interface ISkillForModal {
   id: number;
   name: string;
   category: string;
